+++ src/main/java/com/kerrrusha/fastfood/util/TimeInterval.java
@@ -1,4 +1,4 @@
 package com.kerrrusha.fastfood.util;
 
-public record TimeInterval(int time, int delta) {
+public record TimeInterval(double time, int delta) {
 }
